app/requirements-freeze.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-appconf==1.0.4
django-cors-headers==3.3.0
django-imagekit==4.0.2
django-rest-auth==0.9.5
django-sslserver==0.22
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.18.4
oauthlib==3.1.0
pilkit==2.0
Pillow==7.1.2
proglog==0.1.9
psycopg2==2.8.5
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tqdm==4.46.0
urllib3==1.25.9
uWSGI==2.0.18
